Fix RPM creation; handle non-systemd RPM installations (e.g. Docker)

+++ icecp-node/packaging/scripts/before-install.sh
@@ -6,13 +6,19 @@ before_install()
 {
   echo "Performing pre-install steps for <%= name %> version=<%= version %> iteration=<%= iteration %>"
 
-  # Stop (if any) previous versions of running icecp service
-  if systemctl list-units --type=service | grep 'icecp' >/dev/null; then
-    systemctl stop icecp
-  fi
+  if [ $(cat /proc/1/comm) = "systemd" ]; then
+    echo "Found systemd; attempting to stop any prior icecp service"
 
-  # Disable (if any) previous version of icecp service and remove sym links created by enable command
-  systemctl disable icecp
+    # Stop (if any) previous versions of running icecp service
+    if systemctl list-units --type=service | grep 'icecp' >/dev/null; then
+      systemctl stop icecp
+    fi
+
+    # Disable (if any) previous version of icecp service and remove sym links created by enable command
+    if [ -f $SYSCONFIGDIR/systemd/system/icecp.service ]; then
+      systemctl disable icecp
+    fi
+  fi
 }
 
 before_install
