+++ icecp-node/packaging/scripts/before-remove.sh
@@ -5,14 +5,16 @@ set -e
 before_remove() 
 {
   echo "Performing pre-removal steps for <%= name %> version=<%= version %> iteration=<%= iteration %>"
-  
-  # Stop icecp service
-  if systemctl list-units --type=service | grep 'icecp' >/dev/null; then
-    systemctl stop icecp
-  fi
 
-  # Disable icecp service and remove sym links created by enable command
-  systemctl disable icecp
+  if [ $(cat /proc/1/comm) = "systemd" ]; then
+    # Stop icecp service
+    if systemctl list-units --type=service | grep 'icecp' >/dev/null; then
+    systemctl stop icecp
+    fi
+
+    # Disable icecp service and remove sym links created by enable command
+    systemctl disable icecp
+  fi
 }
 
 before_remove
